Trait syncstorage::db::mysql::models::_impl_queryable_by_name_for_usercollectionsresult::diesel::query_source::joins::JoinTo

pub trait JoinTo<T> {
    type FromClause;
    type OnClause;
    fn join_target(rhs: T) -> (Self::FromClause, Self::OnClause);
}

Indicates that two tables can be joined without an explicit ON clause.

Implementations of this trait are generated by invoking joinable!. Implementing this trait means that you can call left_table.inner_join(right_table), without supplying the ON clause explicitly. To join two tables which do not implement this trait, you will need to call .on.

See joinable! and inner_join for usage examples.

Associated Types

type FromClause

type OnClause

Required Methods

fn join_target(rhs: T) -> (Self::FromClause, Self::OnClause)