Trait syncstorage::db::mysql::schema::user_collections::_impl_query_id_for_table::diesel::query_dsl::GroupByDsl

pub trait GroupByDsl<Expr> where
    Expr: Expression, {
    type Output: Query;
    fn group_by(self, expr: Expr) -> Self::Output;
}

This trait is not yet part of Diesel's public API. It may change in the future without a major version bump.

This trait exists as a stop-gap for users who need to use GROUP BY in their queries, so that they are not forced to drop entirely to raw SQL. The arguments to group_by are not checked, nor is the select statement forced to be valid.

Since Diesel otherwise assumes that you have no GROUP BY clause (which would mean that mixing an aggregate and non aggregate expression in the same query is an error), you may need to use sql for your select clause.

Associated Types

type Output: Query

The type returned by .group_by

Loading content...
Required methods

fn group_by(self, expr: Expr) -> Self::Output

See the trait documentation.

Loading content...