Trait syncstorage::db::results::_impl_queryable_by_name_for_getbso::diesel::expression_methods::EqAll

pub trait EqAll<Rhs> where
    <Self::Output as Expression>::SqlType == Bool, {
    type Output: Expression;
    fn eq_all(self, rhs: Rhs) -> Self::Output;
}

This method is used by FindDsl to work with tuples. Because we cannot express this without specialization or overlapping impls, it is brute force implemented on columns in the column! macro.

Associated Types

type Output: Expression

Required Methods

fn eq_all(self, rhs: Rhs) -> Self::Output