Struct syncstorage::db::util::SyncTimestamp

pub struct SyncTimestamp(u64);

Sync Timestamp

Internally represents a Sync timestamp as a u64 representing milliseconds since the epoch.

Methods

impl SyncTimestamp

pub fn as_header(self) -> String

Create a string value compatible with existing Sync Timestamp headers

Represents the timestamp as second since epoch with two decimal places of precision.

pub fn from_header(val: &str) -> Result<Self, &'static str>

Create a SyncTimestamp from a string header

Assumes the string represents the seconds since epoch with two decimal places of precision.

pub(super) fn from_i64(
val: i64
) -> Result<Self, DbError>

Create a SyncTimestamp from an i64

Only called from the db module

pub fn from_milliseconds(val: u64) -> Self

Create a SyncTimestamp from the milliseconds since epoch

pub fn from_seconds(val: f64) -> Self

Create a SyncTimestamp from seconds since epoch

pub fn as_i64(self) -> i64

Return the timestamp as an i64 milliseconds since epoch

pub fn as_seconds(self) -> f64

Return the timestamp as an f64 seconds since epoch