Struct syncstorage::web::extractors::RequestValidationErrors

pub struct RequestValidationErrors(ValidationErrors);

Request Validation Errors

This is a wrapper on validator's ValidationErrors so that the error can be properly adapted to a ResponseError.

Methods

impl RequestValidationErrors

fn with_location(
errors: ValidationErrors,
location: RequestErrorLocation
) -> Self

Map all the errors to a specific location

This is useful for map_err to map the ValidationErrors from a single location to include the appropriate RequestErrorLocation

fn with_error(
field_name: &'static str,
message: &'static str,
location: RequestErrorLocation
) -> Self

Helper to quickly make a RequestValidationErrors with a single error