Struct syncstorage::db::mysql::models::MysqlDb

pub struct MysqlDb {
    pub(super) conn: PooledConnection<ConnectionManager<MysqlConnection>>,
    pub(super) session: MysqlDbSession,
    coll_cache: Arc<CollectionCache>,
}

Fields
conn: PooledConnection<ConnectionManager<MysqlConnection>> session: MysqlDbSession coll_cache: Arc<CollectionCache>

Pool level cache of collection_ids and their names

Methods

impl MysqlDb

pub fn new(
conn: PooledConnection<ConnectionManager<MysqlConnection>>,
coll_cache: Arc<CollectionCache>
) -> Self

pub fn lock_for_read(
&mut self,
user_id: HawkIdentifier,
collection: &str
) -> Result<()>

APIs for collection-level locking

Explicitly lock the matching row in the user_collections table. Read locks do SELECT ... LOCK IN SHARE MODE and write locks do SELECT ... FOR UPDATE.

In theory it would be possible to use serializable transactions rather than explicit locking, but our ops team have expressed concerns about the efficiency of that approach at scale.

pub fn lock_for_write(
&mut self,
user_id: HawkIdentifier,
collection: &str
) -> Result<()>

pub fn delete_storage_sync(&self, user_id: u32) -> Result<()>

pub fn delete_collection_sync(
&self,
user_id: u32,
collection: &str
) -> Result<i64>

pub(super) fn create_collection(
&self,
name: &str
) -> Result<i32>

fn get_or_create_collection_id(&self, name: &str) -> Result<i32>

pub(super) fn get_collection_id(
&self,
name: &str
) -> Result<i32>

fn _get_collection_name(&self, id: i32) -> Result<String>

pub fn put_bso_sync(&self, bso: &PutBso) -> Result<PutBso>

pub fn get_bsos_sync(
&self,
user_id: u32,
collection: &str,
ids: &[&str],
older: u64,
newer: u64,
sort: Sorting,
limit: i64,
offset: i64
) -> Result<BSOs>

pub fn get_bso_sync(&self, params: &GetBso) -> Result<Option<GetBso>>

pub fn delete_bso_sync(
&self,
user_id: u32,
collection: &str,
bso_id: &str
) -> Result<i64>

pub fn delete_bsos_sync(
&self,
user_id: u32,
collection: &str,
bso_id: &[&str]
) -> Result<i64>

pub fn post_bsos_sync(&self, input: &PostCollection) -> Result<PostCollection>

pub fn get_storage_modified_sync(&self, user_id: u32) -> Result<i64>

pub fn get_collection_modified_sync(
&self,
user_id: u32,
collection: &str
) -> Result<i64>

pub fn get_bso_modified_sync(
&self,
user_id: u32,
collection: &str,
bso_id: &str
) -> Result<i64>

pub fn get_collections_modified_sync(
&self,
params: &GetCollections
) -> Result<GetCollections>

fn map_collection_names<T>(
&self,
by_id: HashMap<i32, T>
) -> Result<HashMap<String, T>>

fn load_collection_names(
&self,
collection_ids: &[i32]
) -> Result<HashMap<i32, String>>

pub(super) fn touch_collection(
&self,
user_id: u32,
collection_id: i32
) -> Result<i64>

pub fn get_storage_size_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetStorageUsage>

pub fn get_collection_sizes_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetCollectionCounts>

pub fn get_collection_counts_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetCollectionCounts>