Struct syncstorage::db::mysql::models::MysqlDb

pub struct MysqlDb {
    pub(super) inner: Arc<MysqlDbInner>,
    coll_cache: Arc<CollectionCache>,
}

Fields
inner: Arc<MysqlDbInner>

Synchronous Diesel calls are executed in a tokio ThreadPool to satisfy the Db trait's asynchronous interface.

Arc provides a Clone impl utilized for safely moving to the thread pool but does not provide Send as the underlying db conn. structs are !Sync (Arc requires both for Send). See the Send impl below.

coll_cache: Arc<CollectionCache>

Pool level cache of collection_ids and their names

Methods

impl MysqlDb

pub fn new(
conn: PooledConnection<ConnectionManager<MysqlConnection>>,
thread_pool: Arc<ThreadPool>,
coll_cache: Arc<CollectionCache>
) -> Self

pub fn lock_for_read_sync(&self, params: LockCollection) -> Result<()>

APIs for collection-level locking

Explicitly lock the matching row in the user_collections table. Read locks do SELECT ... LOCK IN SHARE MODE and write locks do SELECT ... FOR UPDATE.

In theory it would be possible to use serializable transactions rather than explicit locking, but our ops team have expressed concerns about the efficiency of that approach at scale.

pub fn lock_for_write_sync(&self, params: LockCollection) -> Result<()>

pub(super) fn begin(
&self
) -> Result<()>

pub fn commit_sync(&self) -> Result<()>

pub fn rollback_sync(&self) -> Result<()>

pub fn delete_storage_sync(&self, user_id: HawkIdentifier) -> Result<()>

pub fn delete_collection_sync(
&self,
params: DeleteCollection
) -> Result<SyncTimestamp>

pub(super) fn create_collection(
&self,
name: &str
) -> Result<i32>

fn get_or_create_collection_id(&self, name: &str) -> Result<i32>

pub(super) fn get_collection_id(
&self,
name: &str
) -> Result<i32>

fn _get_collection_name(&self, id: i32) -> Result<String>

pub fn put_bso_sync(&self, bso: PutBso) -> Result<PutBso>

pub fn get_bsos_sync(&self, params: GetBsos) -> Result<GetBsos>

pub fn get_bso_ids_sync(&self, params: GetBsos) -> Result<GetBsoIds>

pub fn get_bso_sync(&self, params: GetBso) -> Result<Option<GetBso>>

pub fn delete_bso_sync(&self, params: DeleteBso) -> Result<DeleteBso>

pub fn delete_bsos_sync(&self, params: DeleteBsos) -> Result<DeleteBsos>

pub fn post_bsos_sync(&self, input: PostBsos) -> Result<PostBsos>

pub fn get_storage_timestamp_sync(
&self,
user_id: HawkIdentifier
) -> Result<SyncTimestamp>

pub fn get_collection_timestamp_sync(
&self,
params: GetCollectionTimestamp
) -> Result<SyncTimestamp>

pub fn get_bso_timestamp_sync(
&self,
params: GetBsoTimestamp
) -> Result<SyncTimestamp>

pub fn get_collection_timestamps_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetCollectionTimestamps>

fn map_collection_names<T>(
&self,
by_id: HashMap<i32, T>
) -> Result<HashMap<String, T>>

fn load_collection_names<'a>(
&self,
collection_ids: impl Iterator<Item = &'a i32>
) -> Result<HashMap<i32, String>>

pub(super) fn touch_collection(
&self,
user_id: u32,
collection_id: i32
) -> Result<SyncTimestamp>

pub fn get_storage_usage_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetStorageUsage>

pub fn get_collection_usage_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetCollectionUsage>

pub fn get_collection_counts_sync(
&self,
user_id: HawkIdentifier
) -> Result<GetCollectionCounts>

pub fn create_batch_sync(&self, params: CreateBatch) -> Result<CreateBatch>

pub fn validate_batch_sync(
&self,
params: ValidateBatch
) -> Result<ValidateBatch>

pub fn append_to_batch_sync(
&self,
params: AppendToBatch
) -> Result<AppendToBatch>

pub fn delete_batch_sync(&self, params: DeleteBatch) -> Result<DeleteBatch>

pub fn commit_batch_sync(&self, params: CommitBatch) -> Result<CommitBatch>

pub fn get_batch_sync(&self, params: GetBatch) -> Result<Option<GetBatch>>

pub fn timestamp(&self) -> SyncTimestamp