Module syncstorage::web::extractors

Request header/body/query extractors

Handles ensuring the header's, body, and query parameters are correct, extraction to relevant types, and failing correctly with the appropriate errors if issues arise.

Structs

BatchBsoBody

BsoBodies

BsoBody

BsoParam
Bso id parameter extractor

BsoPutRequest
BSO Request Put extractor

BsoQueryParams
Validator to extract BSO search parameters from the query string.

BsoRequest
BSO Request Delete/Get extractor

CollectionParam
Collection parameter extractor

CollectionPostRequest
Collection Request Post extractor

CollectionRequest
Collection Request Delete/Get extractor

HawkIdentifier
Extract a user-identifier from the authentication token and validate against the URL

KNOWN_BAD_PAYLOAD_REGEX

MetaRequest
Information Requests extractor

UidParam

VALID_COLLECTION_ID_REGEX

VALID_ID_REGEX

Enums

PreConditionHeader
PreCondition Header

RequestErrorLocation
Validation Error Location in the request

Constants

BATCH_MAX_BYTES

BATCH_MAX_IDS

BATCH_MAX_RECORDS

BSO_MAX_PAYLOAD_SIZE

BSO_MAX_SORTINDEX_VALUE

BSO_MAX_TTL

BSO_MIN_SORTINDEX_VALUE

_IMPL_DESERIALIZE_FOR_BatchBsoBody

_IMPL_DESERIALIZE_FOR_BsoBodies

_IMPL_DESERIALIZE_FOR_BsoBody

_IMPL_DESERIALIZE_FOR_BsoParam

_IMPL_DESERIALIZE_FOR_BsoQueryParams

_IMPL_DESERIALIZE_FOR_CollectionParam

_IMPL_DESERIALIZE_FOR_HawkIdentifier

_IMPL_DESERIALIZE_FOR_RequestErrorLocation

_IMPL_DESERIALIZE_FOR_UidParam

_IMPL_SERIALIZE_FOR_BatchBsoBody

_IMPL_SERIALIZE_FOR_BsoBody

_IMPL_SERIALIZE_FOR_HawkIdentifier

_IMPL_SERIALIZE_FOR_RequestErrorLocation

Statics

KNOWN_BAD_PAYLOAD_REGEX

VALID_COLLECTION_ID_REGEX

VALID_ID_REGEX

Functions

deserialize_comma_sep_string
Deserialize a comma separated string

deserialize_present_value
Deserialize a value as True if it exists, False otherwise

deserialize_sync_timestamp
Deserialize a header string value (epoch seconds with 2 decimal places) as SyncTimestamp

request_error
Convenience function to create a ValidationError with additional context

validate_body_bso_id
Verifies the BSO id string is valid

validate_body_bso_payload
Verifies the BSO payload size is valid

validate_body_bso_sortindex
Verifies the BSO sortindex is in the valid range

validate_body_bso_ttl
Verifies the BSO ttl is valid

validate_qs_ids
Verifies that the list of id's is not too long and that the ids are valid